Add files via upload
One way to understand how a city government works is by looking at who it employs and how its employees are compensated. This data contains the names, job title, and compensation for San Francisco city employees on an annual basis from 2011 to 2014.
The original source for this data is https://www.kaggle.com/datasets/kaggle/sf-salaries.  I've taken the raw files here and combined/normalized them into a single CSV file as well as a SQLite database with an equivalently-defined table.